Apply refactor to dependant classes
