+++ core/src/main/java/de/pxav/kelp/core/KelpPlugin.java
@@ -9,8 +9,8 @@ import de.pxav.kelp.core.command.KelpCommandRepository;
 import de.pxav.kelp.core.configuration.ConfigurationRepository;
 import de.pxav.kelp.core.configuration.internal.KelpDefaultConfiguration;
 import de.pxav.kelp.core.inventory.KelpInventoryRepository;
-import de.pxav.kelp.core.listener.EventHandlerRegistration;
-import de.pxav.kelp.core.listener.KelpEventRepository;
+import de.pxav.kelp.core.event.listener.EventHandlerRegistration;
+import de.pxav.kelp.core.event.listener.KelpEventRepository;
 import de.pxav.kelp.core.npc.KelpNpcRepository;
 import de.pxav.kelp.core.particle.effect.ParticleEffectRepository;
 import de.pxav.kelp.core.particle.type.ParticleTypeVersionTemplate;
